refactor: add CalculateLotto, LottoMachine, LottoRank, Validate
파일 분리
당첨 여부 계산 로직
로또 생성 부분
enum 로또 등수 정보
유효성검사
