+++ src/main/java/lotto/LottoMachine.java
@@ -1,0 +1,21 @@
+package lotto;
+
+import camp.nextstep.edu.missionutils.Randoms;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public class LottoMachine {
+
+    protected Lotto makeNumber() {
+        List<Integer> numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6);
+        Lotto lotto = new Lotto(sortAscNumber(numbers));
+        return lotto;
+    }
+
+    private List<Integer> sortAscNumber(List<Integer> numbers) {
+        List<Integer> copyNumbers = new ArrayList<>(numbers);
+        Collections.sort(copyNumbers);
+        return copyNumbers;
+    }
+}
